bmo-assistant/services/voice-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-speech==2.25.0
google-cloud-texttospeech==2.16.0
pydantic==2.6.0
python-multipart==0.0.6
